pyproject.toml
[project]
name = "finance-dashboard-backend"
version = "1.0.0"
description = "A production-ready FastAPI backend for a finance dashboard"
requires-python = ">=3.10"
dependencies = [
    "fastapi==0.115.0",
    "uvicorn[standard]==0.30.0",
    "sqlalchemy==2.0.35",
    "aiosqlite==0.20.0",
    "python-jose[cryptography]==3.3.0",
    "passlib[bcrypt]==1.7.4",
    "bcrypt==4.0.1",
    "python-multipart==0.0.9",
    "python-dotenv==1.0.1",
    "slowapi==0.1.9",
    "pydantic[email]==2.9.0",
    "pydantic-settings==2.5.0",
    "greenlet==3.3.2",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
